ai-service/requirements.txt
fastapi
uvicorn
pillow
numpy
opencv-python
python-multipart
diffusers
torch
torchvision
transformers
accelerate